+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Do not expose `AuthorizationResource` anymore via REST API (always identify by resource class and identifier)
 - Do not explicitly register resources anymore. Create `AuthorizationResource` implicitly when adding a resource action grant
 for a resource that does not yet exist.
+- Allow adding and removing grant inheritance between resources
 
 ## v0.5.4